+++ src/main/java/com/atlauncher/data/minecraft/LibraryTypeAdapter.java
@@ -39,6 +39,10 @@ public class LibraryTypeAdapter implements JsonDeserializer<Library> {
             library = new Gson().fromJson(object, ForgeLibrary.class);
         } else {
             library = new Gson().fromJson(object, Library.class);
+
+            if (library.name.contains("forge") && library.downloads.artifact.url.endsWith("-launcher.jar")) {
+                library.downloads.artifact.url = library.downloads.artifact.url.replace("-launcher.jar", ".jar");
+            }
         }
 
         return library;